add txid tree

+++ src/indexer/syncer.rs
@@ -19,7 +19,7 @@ pub struct Operation {
     pub nullifiers: Vec<Fr>,
     pub commitment_hashes: Vec<Fr>,
     pub bound_params_hash: Fr,
-    pub utxo_batch_tree_number: u16,
+    pub utxo_batch_tree_number: u32,
     pub utxo_batch_start_index: u32,
 }
